Added error middleware to server and converted authentication route to middleware
