+++ app/middleware/error.middleware.js
@@ -1,0 +1,7 @@
+'use strict';
+
+// middleware
+exports.errorFilter = function(err, req, res, next) {
+    console.error(err.stack);
+    res.status(500).send('Internal server error.');
+};
