+++ app/server.js
@@ -4,11 +4,13 @@ var request = require('request');
 var bodyParser = require('body-parser');
 var morgan = require('morgan');
 
-var authenticationRoutes = require('./routes/authentication.routes')(express.Router());
 var attributeRoutes = require('./routes/attribute.routes')(express.Router());
 var userRoutes = require('./routes/user.routes')(express.Router());
 var roleRoutes = require('./routes/role.routes')(express.Router());
 var accessRoutes = require('./routes/access.routes')(express.Router());
+
+var errorMiddleware = require('./middleware/error.middleware');
+var authenticationMiddleware = require('./middleware/authentication.middleware');
 
 var config = require('./config/config');
 
@@ -28,13 +30,16 @@ app.use(bodyParser.json());
 // logging
 app.use(morgan('dev'));
 
-// ROUTES
+// ROUTES & MIDDLEWARE
 // ============================================================================
-app.use('/api/', authenticationRoutes);
+app.use(authenticationMiddleware.authentication);
+
 app.use('/api/attribute', attributeRoutes);
 app.use('/api/user', userRoutes);
 app.use('/api/role', roleRoutes);
 app.use('/api/access', accessRoutes);
+
+app.use(errorMiddleware.errorFilter);
 
 // for debugging
 app.get('/kalle', function(req, res) {
@@ -42,6 +47,11 @@ app.get('/kalle', function(req, res) {
 
     res.send('Your email is: ' + req.headers['x-forwarded-email'] + ' and your accname: ' + req.headers['x-forwarded-user']);
 });
+
+// app.use(function(err, req, res, next) {
+//   console.error(err.stack);
+//   res.status(500).send('Something broke!');
+// });
 
 var server = app.listen(port, function() {
     console.log('Server started: http://localhost:%s/', server.address().port);
